refactoring outputwriter
