+++ phenex/util/output_concatenator/__init__.py
@@ -1,0 +1,4 @@
+from .cohort_group import CohortGroup
+from .output_concatenator import OutputConcatenator
+
+__all__ = ["OutputConcatenator", "CohortGroup"]
